requirements-dev.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==18.9.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
ndg-httpsclient==0.5.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
urllib3==2.0.2
